+++ web/src/components/ConnectingView.module.css
@@ -16,7 +16,6 @@
   background-clip: text;
   -webkit-background-clip: text;
   -webkit-text-fill-color: transparent;
-  font-family: Inter;
   font-size: 20px;
   font-style: normal;
   font-weight: 700;
@@ -26,7 +25,6 @@
 
 .prompt {
   color: #fff;
-  font-family: Inter;
   font-size: 30px;
   font-style: normal;
   font-weight: 700;
@@ -37,7 +35,6 @@
   width: 402px;
   color: #bfbfbf;
   text-align: center;
-  font-family: Inter;
   font-size: 14px;
   font-style: normal;
   font-weight: 500;
