+++ web/src/components/GlobalStyles.module.css
@@ -4,7 +4,6 @@
     justify-content: center;
     align-items: center;
     gap: 5px;
-    /*align-self: stretch;*/
     border: 0.5px solid #f1e3bf;
     background: #00a341;
     color: #fff;
@@ -45,7 +44,6 @@
     -webkit-text-fill-color: transparent;
     color: #FFF;
     text-align: center;
-    font-family: Inter;
     font-size: 24px;
     font-style: italic;
     font-weight: 700;
@@ -64,7 +62,7 @@
         padding: 8px 12px;
         border-radius: 4px;
         border: 1px solid #ccc;
-        box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.2);
+        box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2);
         font-size: 0.9rem;
     }
