+++ web/src/components/Playing.module.css
@@ -1,21 +1,7 @@
 .container {
-  /*display: flex;*/
-  /*width: 648px;*/
-  /*padding: 40px 20px;*/
-  /*flex-direction: column;*/
-  /*justify-content: center;*/
-  /*align-items: center;*/
-  /*width: 100%;*/
-  /*gap: 40px;*/
-  /*!*flex: 1 0 0;*!*/
-  /*border: 1px solid #4d4d4d;*/
-  /*background: rgba(41, 41, 41, 0.7);*/
     display: flex;
-    /*width: 1440px;*/
-    /*height: 800px;*/
     flex-direction: column;
     align-items: flex-start;
-    /*background-color: #221111;*/
 }
 
 .title {
@@ -23,7 +9,6 @@
   background-clip: text;
   -webkit-background-clip: text;
   -webkit-text-fill-color: transparent;
-  font-family: Inter;
   font-size: 20px;
   font-style: normal;
   font-weight: 700;
@@ -33,7 +18,6 @@
 
 .prompt {
   color: #fff;
-  font-family: Inter;
   font-size: 30px;
   font-style: normal;
   font-weight: 700;
