+++ web/src/components/TitleScreen.tsx
@@ -1,32 +1,17 @@
-import { useContext, useEffect, useState } from "react";
+import { useContext, useEffect } from "react";
+import { useQueryClient } from "react-query";
+
 import ConnectingView from "./ConnectingView";
 import Web3Context from "../contexts/Web3Context/context";
 import TitleScreenLayout from "./layout/TitleScreenLayout";
 import PlayingLayout from "./layout/PlayingLayout";
 import Playing from "./Playing";
-import { useRouter } from "next/router";
-import { useQuery, useQueryClient } from "react-query";
-// eslint-disable-next-line @typescript-eslint/no-var-requires
-import queryCacheProps from "../hooks/hookCommon";
 import { useGameContext } from "../contexts/GameContext";
-import OwnedTokens from "./tokens/OwnedTokens";
-import SessionsView from "./sessions/SessionsView";
-// eslint-disable-next-line @typescript-eslint/no-var-requires
 
 const TitleScreen = () => {
-  const [step, setStep] = useState(1);
   const web3ctx = useContext(Web3Context);
 
-  const router = useRouter();
-
-  const { selectedSession, contractAddress, selectedToken, updateContext, chainId } =
-    useGameContext();
-
-  // useEffect(() => {
-  //   if (typeof router.query.session_id === "string") {
-  //     updateContext({ sessionId: Number(router.query.session_id) });
-  //   }
-  // }, [router.query.session_id]);
+  const { chainId } = useGameContext();
 
   const queryClient = useQueryClient();
 
@@ -39,7 +24,7 @@ const TitleScreen = () => {
     <>
       {web3ctx.buttonText !== "Connected" || web3ctx.chainId !== chainId ? (
         <TitleScreenLayout>
-          <ConnectingView nextStep={() => setStep(3)} />
+          <ConnectingView />
         </TitleScreenLayout>
       ) : (
         <>
@@ -53,4 +38,3 @@ const TitleScreen = () => {
 };
 
 export default TitleScreen;
-1;
