+++ web/src/types.d.ts
@@ -153,3 +153,8 @@ interface SwingLocation {
   count: number;
   swing_type: number;
 }
+
+interface EthereumError {
+  code: number;
+  message: string;
+}
