+++ web/src/utils/messages.ts
@@ -1,5 +1,4 @@
 import { Session } from "../types";
-import { SECOND_REVEAL_PRICE_MULTIPLIER } from "../constants";
 
 export const progressMessage = (session: Session) => {
   if (session.progress === 1) {
@@ -44,12 +43,4 @@ export const progressMessage = (session: Session) => {
     }
     return `Strike three! ${session.pair.pitcher?.name} wins it for the visitors `;
   }
-};
-
-export const sendTransactionWithEstimate = async (account: string, method: any) => {
-  const estimatedGas = await method.estimateGas({ from: account });
-  return method.send({
-    from: account,
-    gas: Math.ceil(SECOND_REVEAL_PRICE_MULTIPLIER * estimatedGas),
-  });
 };
